Refactor terraform fixtures together
