+++ harvester_e2e_tests/fixtures/terraform.py
@@ -24,6 +24,137 @@ provider "harvester" {
 }
 '''
 
+TF_PROVIDER_RANCHER = '''
+terraform {
+  required_providers {
+    rancher2 = {
+      source  = "%(provider_source)s"
+      version = "%(provider_version)s"
+    }
+  }
+}
+
+provider "rancher2" {
+  api_url   = "%(rancher_endpoint)s"
+  token_key = "%(rancher_token)s"
+  insecure = true
+}
+
+data "rancher2_cluster_v2" "%(harvester_name)s" {
+  name = "%(harvester_name)s"
+}
+
+'''
+
+# rancher2_machine_config_v2
+TF_MACHINE_CONFIG = '''
+resource "rancher2_machine_config_v2" "%(name)s" {
+  generate_name = "%(name)s"
+  %(harvester_config)s
+}
+
+'''
+TF_HARVESTER_CONFIG = '''
+  harvester_config {
+    vm_namespace = "default"
+    cpu_count = "2"
+    memory_size = "4"
+    ssh_user = "%(ssh_user)s"
+    disk_info = %(disk_info)s
+    network_info = %(network_info)s
+    user_data = %(user_data)s
+  }
+'''
+TF_DISK_INFO = '''<<EOF
+    {
+      "disks": [{
+        "imageName": "%(image_name)s",
+        "size": 40,
+        "bootOrder": 1
+      }]
+    }
+    EOF
+'''
+TF_NETWORK_INFO = '''<<EOF
+    {
+      "interfaces": [{
+        "networkName": "%(network_name)s"
+      }]
+    }
+    EOF
+'''
+TF_USER_DATA = '''<<EOF
+    #cloud-config
+    password: password
+    chpasswd: {expire: False}
+    ssh_pwauth: True
+    package_update: true
+    packages:
+      - qemu-guest-agent
+      - iptables
+    runcmd:
+      - - systemctl
+        - enable
+        - '--now'
+        - qemu-guest-agent.service
+    EOF
+'''
+
+
+# rancher2_cluster_v2
+TF_CLUSTER_CONFIG = '''
+resource "rancher2_cluster_v2" "%(name)s" {
+  name               = "%(name)s"
+  kubernetes_version = "%(rke2_version)s"
+  %(rke_config)s
+}
+
+'''
+
+TF_RKE_CONFIG = '''
+  rke_config {
+    %(machine_pools)s
+    machine_selector_config {
+      config = {
+        cloud-provider-config = file("${path.module}/kubeconfig")
+        cloud-provider-name = "harvester"
+      }
+    }
+    machine_global_config = <<EOF
+    cni: "calico"
+    disable-kube-proxy: false
+    etcd-expose-metrics: false
+    EOF
+    upgrade_strategy {
+      control_plane_concurrency = "10%%"
+      worker_concurrency        = "10%%"
+    }
+    etcd {
+      snapshot_schedule_cron = "0 */5 * * *"
+      snapshot_retention     = 5
+    }
+    chart_values = <<EOF
+    harvester-cloud-provider:
+      clusterName: %(harvester_name)s
+      cloudConfigPath: /var/lib/rancher/rke2/etc/config-files/cloud-provider-config
+    EOF
+}
+'''
+TF_MACHINE_POOLS = '''
+    machine_pools {
+      name                         = "pool1"
+      cloud_credential_secret_name = rancher2_cloud_credential.%(cloud_credential_name)s.id
+      control_plane_role           = true
+      etcd_role                    = true
+      worker_role                  = true
+      quantity                     = 1
+      machine_config {
+        kind = rancher2_machine_config_v2.%(machine_config_name)s.kind
+        name = rancher2_machine_config_v2.%(machine_config_name)s.name
+      }
+    }
+'''
+
 
 def remove_ansicode(ctx):
     if isinstance(ctx, bytes):
@@ -47,6 +178,16 @@ def tf_provider_version(request):
 
 
 @pytest.fixture(scope="session")
+def tf_provider_rancher_ver(request):
+    version = request.config.getoption('--terraform-provider-rancher')
+    if not version:
+        import requests
+        resp = requests.get("https://registry.terraform.io/v1/providers/rancher/rancher2")
+        version = max(resp.json()['versions'], key=parse_version)
+    return version
+
+
+@pytest.fixture(scope="session")
 def tf_executor(tf_script_dir):
     run(str(tf_script_dir / "terraform_install.sh"), stdout=PIPE, stderr=PIPE)
     executor = tf_script_dir / "bin/terraform"
@@ -64,10 +205,29 @@ def tf_harvester(api_client, tf_script_dir, tf_provider_version, tf_executor):
     return harv
 
 
+@pytest.fixture(scope="module")
+def tf_rancher(rancher_api_client, tf_script_dir, tf_provider_rancher_ver, tf_executor,
+               harvester, rancher):
+    tf_rancher = TerraformRancher(tf_executor,
+                                  tf_script_dir / datetime.now().strftime("%Hh%Mm_%m-%d"))
+    kubeconfig = rancher_api_client.generate_kubeconfig(harvester["id"], harvester["name"])
+
+    out, err, exc_code = \
+        tf_rancher.initial_provider(kubeconfig, tf_provider_rancher_ver, harvester, rancher)
+    assert not err and 0 == exc_code
+    return tf_rancher
+
+
 @pytest.fixture(scope="session")
 def tf_resource(tf_provider_version):
     converter = Path("./terraform_test_artifacts/json2hcl")
-    return BaseTerraformResource.for_version(tf_provider_version)(converter)
+    return TerraformResource.for_version(tf_provider_version)(converter)
+
+
+@pytest.fixture(scope="session")
+def tf_rancher_resource(tf_provider_rancher_ver):
+    converter = Path("./terraform_test_artifacts/json2hcl")
+    return TerraformRancherResource.for_version(tf_provider_rancher_ver)(converter)
 
 
 class TerraformHarvester:
@@ -109,6 +269,24 @@ class TerraformHarvester:
 
     def destroy_resource(self, resource_type, resource_name):
         return self.execute(f"destroy -auto-approve -target {resource_type}.{resource_name}")
+
+
+class TerraformRancher(TerraformHarvester):
+    def initial_provider(self, kubeconfig, provider_version, harvester, rancher):
+        kubefile = self.workdir / "kubeconfig"
+        with open(kubefile, "w") as f:
+            f.write(kubeconfig)
+
+        with open(self.workdir / "provider.tf", "w") as f:
+            f.write(TF_PROVIDER_RANCHER % {
+                "provider_source": "rancher/rancher2",
+                "provider_version": provider_version,
+                "rancher_endpoint": rancher["endpoint"],
+                "rancher_token": rancher["token"],
+                "harvester_name": harvester["name"]
+            })
+
+        return self.execute("init")
 
 
 @dataclass
@@ -156,9 +334,10 @@ class BaseTerraformResource:
         if rv.stderr:
             raise TypeError(rv.stderr, rv.stdout, rv.returncode)
         out = rv.stdout.decode()
-        out = re.sub(r'"resource"', "resource", out)  # resource should not quote
-        out = re.sub(r"\"(.+?)\" =", r"\1 =", out)  # property should not quote
-        out = re.sub(r"(.[^ ]+) = {", r"\1 {", out)  # block should not have `=`
+        out = re.sub(r'"resource"', "resource", out)    # resource should not quote
+        out = re.sub(r"\"(.+?)\" =", r"\1 =", out)      # property should not quote
+        out = re.sub(r'"(data\.\S+?)"', r"\1", out)     # data should not quote
+        out = re.sub(r"(.[^ ]+) = {", r"\1 {", out)     # block should not have `=`
         return out
 
     def make_resource(self, resource_type, resource_name, *, convert=True, **properties):
@@ -276,3 +455,54 @@ class TerraformResource_063(TerraformResource_060):
             "harvester_cloudinit_secret", resource_name,
             user_data=user_data, network_data=network_data, convert=convert, **properties
         )
+
+
+class TerraformRancherResource(BaseTerraformResource):
+    ''' https://github.com/rancher/terraform-provider-rancher2/tree/v1.20.0/docs/resources
+    '''
+    support_to = "1.20.0"
+
+    def machine_config(self, rke_cluster_name, network_id, image_id, ssh_user):
+        hcl_str = TF_MACHINE_CONFIG % {
+            "name": rke_cluster_name,
+            "harvester_config": TF_HARVESTER_CONFIG % {
+                "ssh_user": ssh_user,
+                "disk_info": TF_DISK_INFO % {"image_name": image_id},
+                "network_info": TF_NETWORK_INFO % {"network_name": network_id},
+                "user_data": TF_USER_DATA
+            }
+        }
+        return ResourceContext("rancher2_machine_config_v2", rke_cluster_name, hcl_str, "")
+
+    def cluster_config(self, rke_cluster_name, k8s_version, harvester_name, cloud_credential_name):
+        machine_pools = TF_MACHINE_POOLS % {
+            "cloud_credential_name": cloud_credential_name,
+            "machine_config_name": rke_cluster_name
+        }
+        rke_config = TF_RKE_CONFIG % {
+            "machine_pools": machine_pools,
+            "harvester_name": harvester_name
+        }
+        hcl_str = TF_CLUSTER_CONFIG % {
+            "name": rke_cluster_name,
+            "rke2_version": k8s_version,
+            "rke_config": rke_config
+        }
+        return ResourceContext("rancher2_cluster_v2", rke_cluster_name, hcl_str, "")
+
+
+class TerraformRancherResource_123(TerraformRancherResource):
+    ''' https://github.com/rancher/terraform-provider-rancher2/tree/v1.23.0/docs/resources
+    '''
+    support_to = "1.23.0"
+
+    def cloud_credential(self, name, harvester_name, *, convert=True, **properties):
+        harvester_credential_config = {
+            "cluster_id": f"data.rancher2_cluster_v2.{harvester_name}.cluster_v1_id",
+            "cluster_type": "imported",
+            "kubeconfig_content": f"data.rancher2_cluster_v2.{harvester_name}.kube_config"
+        }
+        return self.make_resource("rancher2_cloud_credential", name,
+                                  name=name,
+                                  harvester_credential_config=harvester_credential_config,
+                                  convert=convert, **properties)
